Otra refactorizacion y documentacion del codigo
